Continued benchmark-to-testing migration
Migrated the microbench and replay benchmark runner modules into the testing module, underneath a new bench folder, as these are standalone executable tests
Fixed the runner not compiling due to import/package name changes
