--- benchmarking/runner/ref/comunica/build.gradle.kts
@@ -1,24 +1,0 @@
-plugins {
-    kotlin("multiplatform")
-}
-
-group = "benchmarking.runner"
-
-kotlin {
-    js {
-        nodejs {
-            binaries.library()
-        }
-    }
-
-    sourceSets {
-        val jsMain by getting {
-            dependencies {
-                implementation(project(":benchmarking:runner:core"))
-                implementation(project(":interop:rdfjs"))
-                implementation(npm("@comunica/query-sparql", "4.1.0"))
-                implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.10.1")
-            }
-        }
-    }
-}
