+++ testing/bench/sparql/core/build.gradle.kts
