+++ testing/bench/sparql/ref/blazegraph/gradle.properties
