+++ testing/bench/sparql/ref/blazegraph/src/jvmMain/resources/log4j.properties
