+++ testing/bench/sparql/ref/comunica/src/jsMain/kotlin/dev/tesserakt/benchmarking/ComunicaReference.kt
@@ -1,0 +1,59 @@
+package dev.tesserakt.benchmarking
+
+import dev.tesserakt.interop.rdfjs.n3.N3Store
+import dev.tesserakt.interop.rdfjs.n3.N3Term
+import dev.tesserakt.interop.rdfjs.toN3Triple
+import dev.tesserakt.rdf.types.SnapshotStore
+import kotlinx.coroutines.await
+
+
+class ComunicaReference(private val query: String) : Reference() {
+
+    private val engine = ComunicaQueryEngine()
+    private val store = N3Store()
+    private val settings: dynamic = Any()
+        .apply { asDynamic().sources = arrayOf(store) }
+
+    private var previous = emptyArray<ComunicaBinding>()
+    private var current = emptyArray<ComunicaBinding>()
+    private var checksum = 0
+
+    override fun prepare(diff: SnapshotStore.Diff) {
+        diff.deletions.forEach {
+            store.delete(it.toN3Triple())
+        }
+        diff.insertions.forEach {
+            store.add(it.toN3Triple())
+        }
+    }
+
+    override suspend fun eval() {
+        val query = engine.query(query, settings).await()
+        val results = query.toArray().await()
+        results.forEach { binding ->
+            binding.forEach { term, _ ->
+                checksum += term
+                    .unsafeCast<N3Term>()
+                    .checksumValue
+            }
+        }
+        current = results
+    }
+
+    override fun finish(): Output {
+        val result = compare(current.toList(), previous.toList(), checksum = checksum)
+        checksum = 0
+        previous = current
+        return result
+    }
+
+}
+
+private val N3Term.checksumValue: Int
+    get() = when (termType) {
+        "NamedNode", "Literal", "BlankNode" -> value.length
+
+        "Variable" -> throw IllegalArgumentException("Term `$this` is not supported as a quad term!")
+
+        else -> throw IllegalArgumentException("Unknown term type `$termType`!")
+    }
