+++ testing/bench/sparql/ref/rdfox/build.gradle.kts
@@ -1,0 +1,19 @@
+plugins {
+    kotlin("multiplatform")
+}
+
+group = "benchmarking.runner"
+
+kotlin {
+    jvm()
+
+    sourceSets {
+        val jvmMain by getting {
+            dependencies {
+                implementation(project(":testing:bench:sparql:core"))
+                implementation(files("libs/JRDFox.jar"))
+                implementation(project(":serialization:trig"))
+            }
+        }
+    }
+}
