+++ testing/bench/sparql/src/commonMain/kotlin/dev/tesserakt/benchmarking/OutputObserver.kt
