+++ testing/bench/sparql/src/commonMain/kotlin/dev/tesserakt/benchmarking/TimeObserver.kt
