+++ testing/bench/sparql/src/jsMain/kotlin/dev/tesserakt/benchmarking/RunnerConfig.js.kt
