+++ testing/bench/sparql/src/jsMain/kotlin/dev/tesserakt/benchmarking/Util.js.kt
