+++ testing/bench/sparql/src/jsMain/kotlin/dev/tesserakt/benchmarking/Writer.js.kt
