+++ testing/bench/sparql/src/jvmMain/kotlin/dev/tesserakt/benchmarking/MemoryObserver.kt
