+++ testing/bench/sparql/src/jvmMain/kotlin/dev/tesserakt/benchmarking/Util.jvm.kt
