+++ testing/bench/sparql/src/jvmMain/kotlin/dev/tesserakt/benchmarking/Writer.jvm.kt
